pyproject.toml
[tool.poetry]
name = "aprestaurant-api"
version = "0.1.0"
description = "API for the Aunty Paulet Restaurant App"
authors = ["Oluwaseyifunmi Oyefeso"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
tortoise-orm = "^0.17.6"
python-multipart = "^0.0.5"
cloudinary = "^1.26.0"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^0.19.0"
fastapi-mail = "^0.4.1"
PyJWT = "^2.1.0"
aerich = "^0.5.7"
exponent-server-sdk = "^2.0.0"
pycryptodome = "^3.10.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
